- Add experimental "Extend node"

+++ source/blender/compositor/computing/opencl/COM_OpenCLDevice.h
@@ -46,8 +46,8 @@ class OpenCLDevice : public ComputeDevice {
   int m_compute_units;
   int m_clock_freq;
   int m_global_mem_size;
-  int m_max_img_w;
-  int m_max_img_h;
+  size_t m_max_img_w;
+  size_t m_max_img_h;
   cl_image_format m_supported_formats;
   OneElemImg m_one_elem_imgs[MAX_ONE_ELEM_IMGS];
